move neo4j functions to own package
